Add parameter CBRK_SKIP_PREARM
This parameter is meant to prevent any actuator motion until armed.

When the system is prearmed (i.e. the safety button is pressed, or the
safety button is disabled with CBRK_IO_SAFETY), non-throttle outputs
are active.
This is visible on platforms where servomotors control flight, like
the flaps on a fixed wing or the yaw servo on a tri-copter. If actuator
motion before arming is undesirable, set this parameter.